+++ Archs/ARM/Pool.h
@@ -10,7 +10,7 @@ public:
 	virtual void writeTempData(TempData& tempData) { };
 	virtual void writeSymData(SymbolData& symData);
 private:
-	int RamPos;
+	u64 RamPos;
 	bool armstate;
 };
 
@@ -20,15 +20,15 @@ class ArmPool
 public:
 	ArmPool();
 	void Clear();
-	int AddEntry(int value);
-	int GetCount() { return EntryCount; };
-	int GetEntry(int num) { return Entries[num]; };
-	void SetRamPos(int num) { RamPos = num; };  
-	int GetRamPos() { return RamPos; };
+	u64 AddEntry(u32 value);
+	size_t GetCount() { return EntryCount; };
+	u32 GetEntry(size_t num) { return Entries[num]; };
+	void SetRamPos(u64 num) { RamPos = num; };  
+	u64 GetRamPos() { return RamPos; };
 private:
-	int Entries[512];
-	int EntryCount;
-	int RamPos;
+	u32 Entries[512];
+	size_t EntryCount;
+	u64 RamPos;
 };
 
 class ArmPoolCommand: public CAssemblerCommand
@@ -41,7 +41,7 @@ public:
 	virtual void writeSymData(SymbolData& symData);
 	virtual bool IsPool() { return true; };
 private:
-	int RamPos;
-	int PoolId;
-	int Size;
+	u64 RamPos;
+	size_t PoolId;
+	size_t Size;
 };
