+++ Archs/MIPS/Mips.h
@@ -88,6 +88,6 @@ class MipsElfRelocator: public IElfRelocator
 {
 public:
 	virtual bool relocateOpcode(int type, RelocationData& data);
-	virtual void setSymbolAddress(RelocationData& data, unsigned int symbolAddress, int symbolType);
+	virtual void setSymbolAddress(RelocationData& data, u64 symbolAddress, int symbolType);
 	virtual void writeCtorStub(std::vector<ElfRelocatorCtor>& ctors);
 };
